2009-11-26  Sebastien Pouliot  <sebastien@ximian.com>
	* Random.cs: Fix case where the diff between min and max
	is very large in Next(int,int). Fix bug #558593


svn path=/trunk/mcs/; revision=146993

+++ mcs/class/corlib/System/Random.cs
@@ -122,7 +122,7 @@ namespace System
 			if (diff <= 1)
 				return minValue;
 
-			return (int)(Sample () * diff) + minValue;
+			return (int)((uint)(Sample () * diff) + minValue);
 		}
 
 		public virtual void NextBytes (byte [] buffer)
